backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
langchain==0.3.7
langchain-groq==0.2.1
langchain-core==0.3.15
networkx==3.3
pydantic==2.9.2
httpx==0.27.2
python-multipart==0.0.12
anyio==4.4.0
langchain-google-genai==2.0.1
